requirements.txt
flask
langchain==0.0.267
python-dotenv
PyMuPDF
python-docx 
flask_cors
openai